refactor resource scheduler/offer processing to be an interface
