+++ SingularityBase/src/main/java/com/hubspot/singularity/ResourceSchedulerType.java
@@ -1,0 +1,5 @@
+package com.hubspot.singularity;
+
+public enum ResourceSchedulerType {
+  SINGULARITY
+}
