+++ SingularityService/src/main/docker/singularity.yaml
